Use Product object in CartItem
The Product object needs to be accessible via the Client for the
processOrder function. Using a product ID instead of an object reference
would work fine if we had a way to access the Warehouse object to search
for the Product, however a way to access the Warehouse object is not
currently implemented and probably won't change, so this will change
instead.