Add function to make a payment
The implementation doesn't handle the decimal amounts too well. You
can't make exact payments to two decimal places because of the float
data type being imprecise, and it also allows for smaller-than-a-penny
amounts. It also records these amounts to the transaction with imprecise
values, as well. It works, at least.