Split UI into FSM-based system
The UserInterface class is no longer used. The program is run via
WarehouseContext, which has the main method. The program works by having
the WarehouseContext call the "run()" method on the WarehouseState
subclasses. A subclass of WarehouseState will have its main method logic
in the "run()" method and will return the state that it wishes to
transition to. The "run()" method can return "null" indicating that it
wishes to exit. WarehouseContext will, at this point, determine where to
go after the state returns "null;" it shouldn't always exit immediately.
E.g. When a Clerk exits from a "become a Client" state, it should
return to ClerkMenuState and not actually exit, as per project
requirements.

The "getToken()" and similar helper methods have been moved to static
methods in WarehouseState, so all the subclasses can call them without
issue.

ClerkMenuState, ClientMenuState, and ManagerMenuState have to be
implemented, yet. LoginState was implemented as an example of how to use
the new system.