Modify Makefile for more efficient builds
javac automatically compiles needed class files; it just doesn't update
class files unless explicitly told. The makefile now makes all the
source files the rules and not the class files. A class file really only
needs to be rebuilt when the interface of a used class is changed.
LoginState doesn't use any part of the Money class, but the previous
build process would create a chain of updates up to LoginState, anyway.