Implement half of ClientMenuState
The context won't know what state is what, so the states actually have
to contain the logic knowing where to go and when. I changed the context
accordingly. Editing of the shopping cart and printing the client's
waitlist (which wasn't a requirement in project 1 woohoo get to redesign
the backend so fun) are all that's left to be implemented.